requirements.txt
alembic==1.9.3
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
typing-extensions==4.4.0
uvicorn==0.20.0
zipp==3.13.0
